requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
filelock==3.13.4
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
intel-openmp==2021.4.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
protobuf==4.25.3
pyarrow==16.0.0
pydeck==0.9.0b0
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytube @ git+https://github.com/mokandil/pytube.git@12ac717ab6264e66e142c103ef481d9ec165861d
pytz==2024.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
SpeechRecognition==3.10.3
streamlit==1.33.0
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
